+++ include/gc/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_subdirectory(Dialects)
+add_subdirectory(Dialect)
 add_subdirectory(Transforms)
