+++ include/gc/Dialect/Microkernel/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_mlir_dialect(MicrokernelOps microkernel)
+add_mlir_doc(MicrokernelOps MicrokernelOps gc/Dialect/Microkernel/ -gen-op-doc)
+add_mlir_doc(MicrokernelDialect MicrokernelDialect gc/Dialect/Microkernel/ -gen-dialect-doc)
