+++ include/gc/Dialect/Microkernel/MicrokernelOps.h
@@ -6,9 +6,12 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "gc/Dialects/Microkernel/MicrokernelOps.h"
-#include "gc/Dialects/Microkernel/MicrokernelDialect.h"
-#include "mlir/IR/OpImplementation.h"
+#ifndef GC_DIALECTS_MICROKERNELOPS_H
+#define GC_DIALECTS_MICROKERNELOPS_H
+
+#include "mlir/IR/OpDefinition.h"
 
 #define GET_OP_CLASSES
-#include "gc/Dialects/Microkernel/MicrokernelOps.cpp.inc"
+#include "gc/Dialect/Microkernel/MicrokernelOps.h.inc"
+
+#endif // GC_DIALECTS_MICROKERNELOPS_H
