+++ include/gc/Dialect/OnednnGraph/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_mlir_dialect(OnednnGraphOps onednn_graph)
+add_mlir_doc(OnednnGraphOps OnednnGraphOps gc/Dialect/OnednnGraph/ -gen-op-doc)
+add_mlir_doc(OnednnGraphDialect OnednnGraphDialect gc/Dialect/OnednnGraph/ -gen-dialect-doc)
