--- include/gc/Dialects/Linalgx/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_mlir_dialect(LinalgxOps linalgx)
-add_mlir_doc(LinalgxOps LinalgxOps gc/Dialects/Linalgx/ -gen-op-doc)
-add_mlir_doc(LinalgxDialect LinalgxDialect gc/Dialects/Linalgx/ -gen-dialect-doc)
