--- include/gc/Dialects/OnednnGraph/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_mlir_dialect(OnednnGraphOps onednn_graph)
-add_mlir_doc(OnednnGraphOps OnednnGraphOps gc/Dialects/OnednnGraph/ -gen-op-doc)
-add_mlir_doc(OnednnGraphDialect OnednnGraphDialect gc/Dialects/OnednnGraph/ -gen-dialect-doc)
