+++ lib/gc/Dialect/Linalgx/LinalgxDialect.cpp
@@ -1,4 +1,4 @@
-//===- MicrokernelDialect.h - microkernel dialect ---------------*- C++ -*-===//
+//===- LinalgxDialect.h - linalgx dialect -----------------------*- C++ -*-===//
 //
 // This file is licensed under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,15 +6,15 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "gc/Dialects/Microkernel/MicrokernelDialect.h"
-#include "gc/Dialects/Microkernel/MicrokernelOps.h"
+#include "gc/Dialect/Linalgx/LinalgxDialect.h"
+#include "gc/Dialect/Linalgx/LinalgxOps.h"
 
 using namespace mlir;
-using namespace mlir::microkernel;
+using namespace mlir::linalgx;
 
-void MicrokernelDialect::initialize() {
+void LinalgxDialect::initialize() {
   addOperations<
 #define GET_OP_LIST
-#include "gc/Dialects/Microkernel/MicrokernelOps.cpp.inc"
+#include "gc/Dialect/Linalgx/LinalgxOps.cpp.inc"
       >();
 }
